requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
python-dotenv==1.2.1
python-telegram-bot==13.15
pytz==2025.2
requests==2.32.5
ruff==0.15.0
six==1.17.0
tornado==6.1
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
